Refactor models views (#542)
* Refactor: Move story view code to be consistent

* Refactor: Move story model code to be consistent

* Refactor: Rename user view code to auth

* Refactor: Move story view code to be consistent

* Refactor: Move story model code to be consistent

* Refactor: Rename user view code to auth

* Event pagination (#541)

* initial folly

* nearly working pagination

* working pagination

* merged master, bug fixes

* cleanup test file

* working pagination for real

* limited results to 10k due to es limits

* limited results to 10k due to es limits

* nearly there

* nearly there

* nearly there

* fixed tests

* fixed tests

* datastore comment

* datastore comment

* Update bootstrap.sh

* Refactor: Move story view code to be consistent

* Refactor: Move story model code to be consistent

* Refactor: Rename user view code to auth

* Refactor: Move story view code to be consistent

* Refactor: Move story model code to be consistent

* Add home view test back

* Newline
