--- timesketch/models/story.py
@@ -1,50 +1,0 @@
-# Copyright 2015 Google Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""This module implements the models for Timesketch stories."""
-
-from sqlalchemy import Column
-from sqlalchemy import ForeignKey
-from sqlalchemy import Integer
-from sqlalchemy import Unicode
-from sqlalchemy import UnicodeText
-
-from timesketch.models import BaseModel
-from timesketch.models.acl import AccessControlMixin
-from timesketch.models.annotations import LabelMixin
-from timesketch.models.annotations import CommentMixin
-from timesketch.models.annotations import StatusMixin
-
-
-class Story(AccessControlMixin, LabelMixin, StatusMixin, CommentMixin,
-            BaseModel):
-    """Implements the Story model."""
-    title = Column(Unicode(255))
-    content = Column(UnicodeText())
-    user_id = Column(Integer, ForeignKey(u'user.id'))
-    sketch_id = Column(Integer, ForeignKey(u'sketch.id'))
-
-    def __init__(self, title, content, sketch, user):
-        """Initialize the Story object.
-
-        Args:
-            title: The title of the story
-            content: Content of the story
-            sketch: A sketch (instance of timesketch.models.sketch.Sketch)
-            user: A user (instance of timesketch.models.user.User)
-        """
-        super(Story, self).__init__()
-        self.title = title
-        self.content = content
-        self.sketch = sketch
-        self.user = user
