--- timesketch/models/story_test.py
@@ -1,31 +1,0 @@
-# Copyright 2015 Google Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Tests for the story model."""
-
-from timesketch.lib.testlib import ModelBaseTest
-from timesketch.models.story import Story
-
-
-class StoryModelTest(ModelBaseTest):
-    """Tests the user model."""
-
-    def test_story_model(self):
-        """
-        Test that the test story has the expected data stored in the
-        database.
-        """
-        expected_result = frozenset(
-            [(u'title', u'Test'), (u'content', u'Test'), (u'user', self.user1),
-             (u'sketch', self.sketch1)])
-        self._test_db_object(expected_result=expected_result, model_cls=Story)
